refactor(ke-graphql): extract #shared leaf, break compiler↔leaf cycles
Seams audit found value-level circular dependencies: compiler was both the
orchestrator AND the home of shared constants/types, so dataloader/resolver/
tbox imported values back from #compiler, forming compiler ↔ {dataloader,
resolver,tbox} barrel cycles (init-safe but not a strict DAG).

FIX 1 — new #shared leaf domain (src/lib/shared):
- constants.ts: the cross-domain vocab/scalar/structural constants — RDF,
  RDFS, OWL, XSD, SKOS, SH, RDF_TYPE, RDFS_LABEL, STANDARD_NAMESPACES,
  XSD_SCALARS, RESERVED_TYPE_NAMES, RESERVED_FIELD_NAMES, CONNECTION_ARGS.
  Compiler-only IRIs (RDF_PROPERTY, OWL_*, SH_*, XSD_PATTERN, ARTIFACT_VERSION)
  stay in compiler/constants.ts, now deriving from #shared base namespaces.
- getLocalName.ts: moved (imported cross-domain by dataloader). getNamespace/
  isStandardVocab stay (compiler-only).
- types.ts: the IR/value/diagnostic/context contract graph (RawExtraction +
  sub-types, OntologyIR, ClassNode, PropertyNode, RangeSpec, CardinalitySpec,
  NamespaceInfo, InstanceStats, NameMap, Mapped*, FieldTypeSpec,
  ResolverTemplate, EntityValue, TripleSet, TripleValue, InverseKey,
  CompilerContext, RuntimeWarningHandler, QueryFn, Diagnostic*, PassResult).
  Compiler-API types (SchemaPluginOptions/Api, CompilerResult,
  SerializedExtraction, ContextFactory, CustomMapping(s), NonNullOverrides,
  SchemaExtensions*) stay in compiler/types.ts, importing moved types from
  #shared.
- index.ts barrel (@module shared); #shared added to package.json imports with
  the standard development/types/default conditional shape.
- Every importer rewired to #shared. compiler/index.ts re-exports #shared so
  the public root API (src/lib/index.ts) is byte-identical.

Final production DAG (no cycle; #shared is a leaf importing only graphql/
dataloader/@canonical/ke):
  shared, hardening → dataloader → resolver → tbox, execution → compiler → http

FIX 2 — rename storeQueryFn → createStoreQueryFn (noun→verb factory, matches
sibling createPluginQueryFn). File, export, barrels, all ke-graphql usages
(tests included), and packages/cli/pragma (compileSchema.ts).

FIX 3 — createBoundedCache.get(): @note documenting LRU recency mutation.

FIX 4 — branch-agnostic wording ("HTTP layer"/"consumers") in hardening
index/constants + #shared docs.

Gates: ke-graphql check (biome+tsc+webarchitect) ✓, 128 tests ✓ (coverage
88/75/93/88, unchanged), build ✓, root+http import ✓; cli/pragma check:ts ✓
+ 1065 tests ✓. Public root API delta is exactly storeQueryFn→createStoreQueryFn.
