+++ packages/runtime/ke-graphql/package.json
@@ -41,6 +41,11 @@
       "types": "./dist/types/lib/resolver/index.d.ts",
       "default": "./dist/esm/lib/resolver/index.js"
     },
+    "#shared": {
+      "development": "./src/lib/shared/index.ts",
+      "types": "./dist/types/lib/shared/index.d.ts",
+      "default": "./dist/esm/lib/shared/index.js"
+    },
     "#tbox": {
       "development": "./src/lib/tbox/index.ts",
       "types": "./dist/types/lib/tbox/index.d.ts",
