+++ packages/runtime/ke-graphql/src/lib/dataloader/createListLoader.ts
@@ -12,7 +12,7 @@ import {
   type QueryFn,
   RDF_TYPE,
   RDFS_LABEL,
-} from "#compiler";
+} from "#shared";
 
 /**
  * Resolve the name predicate for a class: a declared property in the class's
