+++ packages/runtime/ke-graphql/src/lib/dataloader/uris.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import type { NamespaceInfo } from "#compiler";
+import type { NamespaceInfo } from "#shared";
 import { toFull, toPrefixed } from "./uris.js";
 
 const namespaces = new Map<string, NamespaceInfo>([
