+++ packages/runtime/ke-graphql/src/lib/execution/incremental.test.ts
@@ -6,12 +6,8 @@
 import { createTestStore } from "@canonical/ke/testing";
 import { parse } from "graphql";
 import { afterEach, describe, expect, it } from "vitest";
-import {
-  type CompilerContext,
-  type CompilerResult,
-  compile,
-  storeQueryFn,
-} from "#compiler";
+import { type CompilerResult, compile, createStoreQueryFn } from "#compiler";
+import type { CompilerContext } from "#shared";
 import { DS_REALISTIC_TTL, MINIMAL_TTL, PREFIXES } from "#testing";
 import extractStatic from "./extractStatic.js";
 import {
@@ -37,7 +33,7 @@ const setup = async (
 ): Promise<{ result: CompilerResult; context: CompilerContext }> => {
   const { store, cleanup } = await createTestStore({ ttl, prefixes: PREFIXES });
   cleanups.push(cleanup);
-  const result = await compile(storeQueryFn(store), PREFIXES, options);
+  const result = await compile(createStoreQueryFn(store), PREFIXES, options);
   return { result, context: result.createContext(store) };
 };
