+++ packages/runtime/ke-graphql/src/lib/resolver/coerce.ts
@@ -3,7 +3,7 @@
 // stripping (EC.06). Runtime failures go to the context's warning channel.
 // =============================================================================
 
-import { type RuntimeWarningHandler, XSD } from "#compiler";
+import { type RuntimeWarningHandler, XSD } from "#shared";
 import type { ScalarName } from "./types.js";
 
 /**
