+++ packages/runtime/ke-graphql/src/lib/shared/getLocalName.ts
